pyproject.toml
[project]
name = "generating-engine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic",
    "pyyaml",
    "dishka",
    "fastapi",
    "uvicorn",
    "langchain",
    "langchain-openai",
    "langchain-core",
    "mcp[cli]>=1.23.1",
    "langgraph>=1.0.4",
]
